pyproject.toml
[project]
name = "toy-langchain"
version = "0.1.0"
description = "LangChain project with Google Gemini"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.23",
    "langchain-google-genai>=0.0.11",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.2",
]
